+++ autoform-events.js
@@ -53,6 +53,7 @@ Template.autoForm.events({
     var currentDoc = data.doc;
     var docId = currentDoc ? currentDoc._id : null;
     var resetOnSuccess = data.resetOnSuccess;
+    var isValid;
 
     // Make sure we have a collection if we need one for the requested submit type
     if (!collection) {
@@ -83,8 +84,25 @@ Template.autoForm.events({
 
     function failedValidation() {
       selectFirstInvalidField(formId, ss, template);
+      var ec = ss.namedContext(formId);
+      var ik = ec.invalidKeys(), err;
+      if (ik) {
+        if (ik.length) {
+          // We add `message` prop to the invalidKeys.
+          // Maybe SS pkg should just add that property back in?
+          ik = _.map(ik, function (o) {
+            return _.extend({message: ec.keyErrorMessage(o.name)}, o);
+          });
+          err = new Error(ik[0].message);
+        } else {
+          err = new Error('form failed validation');
+        }
+        err.invalidKeys = ik;
+      } else {
+        err = new Error('form failed validation');
+      }
       _.each(onError, function onErrorEach(hook) {
-        hook('pre-submit validation', new Error('form failed validation'), template);
+        hook('pre-submit validation', err, template);
       });
       haltSubmission();
     }
@@ -236,30 +254,26 @@ Template.autoForm.events({
       return;
     }
 
-    // Validate, which also gets form values.
+    // Gather all form values
+    var formDocs = getFormValues(template, formId, ss);
+    var insertDoc = formDocs.insertDoc;
+    var updateDoc = formDocs.updateDoc;
+
+    // Pre-validate
     // For inserts and updates, which have their
     // own validation, we validate here only if
     // there is a `schema` attribute on the form.
     // Otherwise we let collection2 do the validation
     // after before hooks have run.
-    var skipValidation;
     if (data.validationType !== 'none' && (ssIsOverride || isMethod || isNormalSubmit)) {
-      skipValidation = false;
-      // For method forms when ssIsOverride, we will validate again, later, after before hooks
+      isValid = _validateForm(formId, data, formDocs);
+      // If we failed pre-submit validation, we stop submission.
+      if (isValid === false) {
+        return failedValidation();
+      }
+      // NOTE: For method forms when ssIsOverride, we will validate again, later, after before hooks
       // but before calling the method, against the collection schema
-    } else {
-      skipValidation = true;
     }
-
-    var results = _validateForm(formId, template, skipValidation);
-
-    // If we failed pre-submit validation, we stop submission.
-    if (results.insertDocIsValid === false) {
-      return failedValidation();
-    }
-
-    var insertDoc = results.insertDoc;
-    var updateDoc = results.updateDoc;
 
     // Run beginSubmit hooks (disable submit button or form, etc.)
     // NOTE: This needs to stay after getFormValues in case a
@@ -283,7 +297,15 @@ Template.autoForm.events({
         // Make callback for insert
         var insertCallback = makeCallback('insert');
         // Perform insert
-        collection.insert(doc, {validationContext: formId}, insertCallback);
+        if (typeof collection.simpleSchema === "function" && collection.simpleSchema() != null) {
+          // If the collection2 pkg is used and a schema is attached, we pass a validationContext
+          collection.insert(doc, {validationContext: formId}, insertCallback);
+        } else {
+          // If the collection2 pkg is not used or no schema is attached, we don't pass options
+          // because core Meteor's `insert` function does not accept
+          // an options argument.
+          collection.insert(doc, insertCallback);
+        }
       });
     }
 
@@ -313,9 +335,9 @@ Template.autoForm.events({
         // When both `schema` and `collection` are supplied, we do a
         // second validation now, against the collection schema,
         // before calling the method.
-        if (ssIsOverride && data.validationType !== 'none') {
-          var isValid = validateFormDoc(doc, false, formId, ss);
-          if (!isValid) {
+        if (ssIsOverride) {
+          isValid = _validateForm(formId, data, formDocs, true);
+          if (isValid === false) {
             return failedValidation();
           }
         }
@@ -349,28 +371,33 @@ Template.autoForm.events({
     }
   },
   'change form': function autoFormChangeHandler(event, template) {
+    var self = this;
+
     var key = event.target.getAttribute("data-schema-key");
     if (!key)
       return;
 
-    var formId = this.id;
+    var formId = self.id || defaultFormId;
     var data = formData[formId];
-    var validationType = 'submitThenKeyup';
+    if (!data)
+      return;
 
-    if (data && data.ss) {
-      var ss = data.ss;
-      formPreserve.registerForm(formId, function autoFormRegFormCallback() {
-        return getFormValues(template, formId, ss).insertDoc;
-      });
+    // Update cached form values for hot code reload persistence
+    formPreserve.registerForm(formId, function autoFormRegFormCallback() {
+      return getFormValues(template, formId, data.ss).insertDoc;
+    });
 
-      // Get field's value for reactive show/hide of other fields by value
-      updateTrackedFieldValue(formId, key, getFieldValue(template, key));
+    // Update field's value for reactive show/hide of other fields by value
+    updateTrackedFieldValue(formId, key, getFieldValue(template, key));
+
+    // If the form should be auto-saved whenever updated, we do that on field
+    // changes instead of validating the field
+    if (data.autosave) {
+      $(event.currentTarget).submit();
+      return;
     }
 
-    if(data){
-      validationType = data.validationType || 'submitThenKeyup';
-    }
-    
+    var validationType = data.validationType || 'submitThenKeyup';
     var onlyIfAlreadyInvalid = (validationType === 'submitThenKeyup' || validationType === 'submitThenBlur');
     if (validationType === 'keyup' || validationType === 'blur' || validationType === 'submitThenKeyup' || validationType === 'submitThenBlur') {
       validateField(key, template, false, onlyIfAlreadyInvalid);